+++ Dtos/InventoryForUpdateDto.cs
@@ -5,7 +5,7 @@ namespace CheckIT.API.Dtos
     public class InventoryForUpdateDto
     {
         public int Id { get; set; }
-		public int UPC { get; set; }
+		public long UPC { get; set; }
 		public decimal Price { get; set; }
         public string Name { get; set; }
 		public string Description { get; set; }
